+++ android-sdk/src/main/java/com/sensorberg/sdk/receivers/SensorbergBroadcastReceiver.java
@@ -1,4 +1,4 @@
-package com.sensorberg.sdk.background;
+package com.sensorberg.sdk.receivers;
 
 import android.content.BroadcastReceiver;
 import android.content.ComponentName;
